Wizard/ImageSourceSelect: modularize on-prem components
Move OnPrem, RegistryAuth, groupByName, and OnPrem.css into an
OnPrem/ directory with a barrel export. ImageSourceError stays at the
parent level since it is shared with Hosted.
